move some chainweb properties into Chainweb.Version
