+++ src/Chainweb/Miner/Miners.hs
@@ -52,7 +52,7 @@ import qualified System.Random.MWC.Distributions as MWC
 -- internal modules
 
 import Chainweb.BlockHeader
-import Chainweb.Difficulty (BlockRate(..), blockRate, encodeHashTarget)
+import Chainweb.Difficulty (encodeHashTarget)
 import Chainweb.Miner.Config (MinerCount(..))
 import Chainweb.Miner.Core
 import Chainweb.RestAPI.Orphans ()
@@ -61,7 +61,7 @@ import Chainweb.RestAPI.Utils
 #endif
 import Chainweb.Time (Seconds(..))
 import Chainweb.Utils (int, partitionEithersNEL, runGet)
-import Chainweb.Version (ChainId, ChainwebVersion(..), order, _chainGraph)
+import Chainweb.Version
 
 ---
