+++ src/Chainweb/Version.hs
@@ -30,8 +30,15 @@ module Chainweb.Version
 , chainwebVersionFromText
 , chainwebVersionToText
 , chainwebVersionId
-, MiningProtocol(..)
-, miningProtocol
+
+-- * Properties of Chainweb Version
+-- ** POW
+, BlockRate(..)
+, blockRate
+, WindowWidth(..)
+, window
+, MinAdjustment(..)
+, minAdjust
 
 -- * Typelevel ChainwebVersion
 , ChainwebVersionT(..)
@@ -98,6 +105,8 @@ import GHC.Generics (Generic)
 import GHC.Stack
 import GHC.TypeLits
 
+import Numeric.Natural
+
 import System.Random
 
 -- internal modules
@@ -106,6 +115,7 @@ import Chainweb.ChainId
 import Chainweb.Crypto.MerkleLog
 import Chainweb.Graph
 import Chainweb.MerkleUniverse
+import Chainweb.Time (Seconds(..))
 import Chainweb.Utils
 
 import Data.Singletons
@@ -298,20 +308,6 @@ instance HasTextRepresentation ChainwebVersion where
     fromText = chainwebVersionFromText
     {-# INLINE fromText #-}
 
--- | Fundamental mining strategies used by the various values of
--- `ChainwebVersion`.
---
-data MiningProtocol = Timed | ProofOfWork
-
-miningProtocol :: ChainwebVersion -> MiningProtocol
-miningProtocol Test{} = Timed
-miningProtocol TimedConsensus{} = Timed
-miningProtocol TimedCPM{} = Timed
-miningProtocol FastTimedCPM{} = Timed
-miningProtocol PowConsensus{} = ProofOfWork
-miningProtocol Development = ProofOfWork
-miningProtocol Testnet02 = ProofOfWork
-
 -- -------------------------------------------------------------------------- --
 -- Value Maps
 
@@ -400,22 +396,6 @@ testVersionToCode Testnet02 =
 fromTestChainwebVersionId :: HasCallStack => Word32 -> ChainwebVersion
 fromTestChainwebVersionId i =
     uncurry ($) . bimap codeToTestVersion (knownGraph . codeToGraph) $ splitTestCode i
-
--- -------------------------------------------------------------------------- --
--- Basic Properties
-
-chainwebVersionGraph :: ChainwebVersion -> ChainGraph
-chainwebVersionGraph (Test g) = g
-chainwebVersionGraph (TimedConsensus g) = g
-chainwebVersionGraph (PowConsensus g) = g
-chainwebVersionGraph (TimedCPM g) = g
-chainwebVersionGraph (FastTimedCPM g) = g
-chainwebVersionGraph Development = petersonChainGraph
-chainwebVersionGraph Testnet02 = petersonChainGraph
-
-instance HasChainGraph ChainwebVersion where
-    _chainGraph = chainwebVersionGraph
-    {-# INLINE _chainGraph #-}
 
 -- -------------------------------------------------------------------------- --
 -- Type level ChainwebVersion
@@ -508,3 +488,87 @@ randomChainId :: HasChainwebVersion v => v -> IO ChainId
 randomChainId v = (!!) (toList cs) <$> randomRIO (0, length cs - 1)
   where
     cs = chainIds v
+
+-- -------------------------------------------------------------------------- --
+-- Properties of Chainweb Versions
+-- -------------------------------------------------------------------------- --
+
+-- -------------------------------------------------------------------------- --
+-- Graph
+
+chainwebVersionGraph :: ChainwebVersion -> ChainGraph
+chainwebVersionGraph (Test g) = g
+chainwebVersionGraph (TimedConsensus g) = g
+chainwebVersionGraph (PowConsensus g) = g
+chainwebVersionGraph (TimedCPM g) = g
+chainwebVersionGraph (FastTimedCPM g) = g
+chainwebVersionGraph Development = petersonChainGraph
+chainwebVersionGraph Testnet02 = petersonChainGraph
+
+instance HasChainGraph ChainwebVersion where
+    _chainGraph = chainwebVersionGraph
+    {-# INLINE _chainGraph #-}
+
+-- -------------------------------------------------------------------------- --
+-- POW Parameters
+
+-- | The gap in SECONDS that we desire between the Creation Time of subsequent
+-- blocks in some chain.
+--
+newtype BlockRate = BlockRate Seconds
+
+-- | The Proof-of-Work `BlockRate` for each `ChainwebVersion`. This is the
+-- number of seconds we expect to pass while a miner mines on various chains,
+-- eventually succeeding on one.
+--
+blockRate :: ChainwebVersion -> Maybe BlockRate
+blockRate Test{} = Nothing
+blockRate TimedConsensus{} = Just $ BlockRate 4
+blockRate PowConsensus{} = Just $ BlockRate 10
+blockRate TimedCPM{} = Just $ BlockRate 4
+-- 120 blocks per hour, 2,880 per day, 20,160 per week, 1,048,320 per year.
+blockRate FastTimedCPM{} = Just $ BlockRate 1
+blockRate Development = Just $ BlockRate 30
+-- 120 blocks per hour, 2,880 per day, 20,160 per week, 1,048,320 per year.
+blockRate Testnet02 = Just $ BlockRate 30
+
+-- | The number of blocks to be mined after a difficulty adjustment, before
+-- considering a further adjustment. Critical for the "epoch-based" adjustment
+-- algorithm seen in `hashTarget`.
+--
+newtype WindowWidth = WindowWidth Natural
+
+-- | The Proof-of-Work `WindowWidth` for each `ChainwebVersion`. For chainwebs
+-- that do not expect to perform POW, this should be `Nothing`.
+--
+window :: ChainwebVersion -> Maybe WindowWidth
+window Test{} = Nothing
+window TimedConsensus{} = Nothing
+-- 5 blocks, should take 50 seconds.
+window PowConsensus{} = Just $ WindowWidth 8
+window TimedCPM{} = Nothing
+window FastTimedCPM{} = Nothing
+-- 120 blocks, should take 1 hour given a 30 second BlockRate.
+window Development = Just $ WindowWidth 120
+-- 120 blocks, should take 1 hour given a 30 second BlockRate.
+window Testnet02 = Just $ WindowWidth 120
+
+-- | The minimum factor of change that a single application of `adjust` must
+-- apply to some `HashTarget` for it to be accepted. As mentioned in `adjust`,
+-- this value should be above \(e = 2.71828\cdots\).
+--
+newtype MinAdjustment = MinAdjustment Natural
+
+-- | The Proof-of-Work `MinAdjustment` for each `ChainwebVersion`. For chainwebs
+-- that do not expect to perform POW, this should be `Nothing`.
+--
+minAdjust :: ChainwebVersion -> Maybe MinAdjustment
+minAdjust Test{} = Nothing
+minAdjust TimedConsensus{} = Nothing
+minAdjust PowConsensus{} = Just $ MinAdjustment 3
+minAdjust TimedCPM{} = Nothing
+minAdjust FastTimedCPM{} = Nothing
+-- See `adjust` for motivation.
+minAdjust Development = Just $ MinAdjustment 3
+minAdjust Testnet02 = Just $ MinAdjustment 3
+
